+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 - Removed: Old deprecated ssl library written in C-Code for better maintenance, stability and smaller file sizes
 - Added: Direct Logging to file without logcat, thanks to @Anton111111
 
-### v.2.3.0-beta
+### v.2.3.0
 - Added some new buttons for keymap
 - Fixed 3 Fatal errors
 - Fixed video decoder settings for allwinner devices
@@ -20,6 +20,8 @@
 - Moved Mic settings to own fragment and added 3 new options for the new mic enhancement from version 2.2.2, which defaults to off for better compatibility
 - Merged PR #481 - Apply MediaTek 60fps and audio optimizations, thanks to @mrkontrast-coder
 - Some rewrite of the AudioTrackWrapper, to enhance stability and minimize stutters
+- Merged PR #490 - Add UI scale settings, thanks to @Anton111111
+- Merged PR #502 - Navigation Broadcast Updates. Thanks to @Bastel2020
 
 ### v.2.2.2
 - Fixed: Exit on disconnect now stops the carmode too